add a test for each route (currently WIP; still gotta write the tests)
