+++ tests/views/test_account.py
@@ -1,0 +1,28 @@
+def test_get_data(client):
+    """Test the get_data route"""
+    # TODO
+    assert True
+
+
+def test_load_schedule(client):
+    """Test the load_schedule(id) route"""
+    # TODO
+    assert True
+
+
+def test_delete_schedule(client):
+    """Test the delete_schedule route"""
+    # TODO
+    assert True
+
+
+def test_update_label(client):
+    """Test the update_label(id) route"""
+    # TODO
+    assert True
+
+
+def test_save(client):
+    """Test the save route"""
+    # TODO
+    assert True
