+++ tests/views/test_classroom.py
@@ -1,0 +1,10 @@
+def test_get_data(client):
+    """Test the get_data route"""
+    # TODO
+    assert True
+
+
+def test_get_occupation(client):
+    """Test the get_occupation(id) route"""
+    # TODO
+    assert True
